Better sepearte query, graph, and store
